FitnessHelper
This is an application that was created to make it easier for people to do fitness. It can calculate the calorie rate for a person depending on the type of diet, choose one of 4 training programs and in the application you can take notes on weight changes. Also , registration and authorization using Email is implemented here , which allows the user to operate with their data from any computer and the server can also handle many clients . There are server and client parts of the program in the repository.